@@ -5,7 +5,7 @@ import '../common.dart';
 
 class NotStickyExample extends StatelessWidget {
   const NotStickyExample({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
@@ -24,11 +24,11 @@ class NotStickyExample extends StatelessWidget {
 
 class _NotStickyList extends StatelessWidget {
   const _NotStickyList({
-    Key key,
+    Key? key,
     this.index,
   }) : super(key: key);
 
-  final int index;
+  final int? index;
 
   @override
   Widget build(BuildContext context) {
